api/requirements.txt
fastapi
uvicorn
openai
boto3
python-dotenv
requests
pillow
scikit-learn
webcolors
